pyproject.toml
[project]
name = "inventory-management-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "inventory>=2.0.1",
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
]
